+++ node/typings/Checkout.ts
@@ -324,3 +324,31 @@ interface ItemWithSimulationInput {
     sellerId: string
   }>
 }
+
+interface RatesAndBenefitsData {
+  rateAndBenefitsIdentifiers: {
+    id: string
+    name: string
+    featured: boolean
+    description: string
+  }[]
+  teaser: {
+    featured: boolean
+    id: string
+    name: string
+    conditions: {
+      parameters: {
+        name: string
+        value: string
+      }[]
+      minimumQuantity: number
+    }
+    effects: {
+      parameters: {
+        name: string
+        value: string
+      }[]
+    }
+    teaserType: string
+  }[]
+}