+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- `teasers` and `discountHighlights` to the `itemsWithSimulation`.
+- `teasers` and `discountHighlights` and payment name to the `itemsWithSimulation`.
 
 ## [2.141.0] - 2021-04-26